--- src/lib/components/pickers/monthselector/index.ts
@@ -1,1 +1,0 @@
-export {default as MonthSelector} from "./MonthSelector.svelte";
