+++ src/lib/components/pickers/monthstepper/MonthStepper.svelte
@@ -7,7 +7,7 @@
     import type {ISizeProperties} from "../../../types/sizes";
     import type {IMarginProperties, IPaddingProperties} from "../../../types/spacings";
 
-    import {get_yearstamp, is_year_in_range} from "../../../util/datetime";
+    import {get_monthstamp, is_month_in_range} from "../../../util/datetime";
     import {BROWSER_CALENDAR, BROWSER_LOCALE} from "../../../util/locale";
 
     import Spacer from "../../layouts/spacer/Spacer.svelte";
@@ -26,7 +26,6 @@
         min?: string;
 
         value: string;
-
         palette?: PROPERTY_PALETTE;
     } & IHTML5Properties &
         IGlobalProperties &
@@ -51,37 +50,50 @@
     export let max: $$Props["max"] = undefined;
     export let min: $$Props["min"] = undefined;
 
-    export let value: $$Props["value"] = get_yearstamp(calendar);
+    export let value: $$Props["value"] = get_monthstamp(calendar);
 
     export let palette: $$Props["palette"] = undefined;
 
-    function on_year_select(difference: number, event: MouseEvent): void {
-        value = _year.add({years: difference}).toString();
+    function on_month_select(difference: number, event: MouseEvent): void {
+        // HACK: Switch to only using `Temporal.PlainYearMonth.add` whenever bug for chained-subtractions is released
+        // https://github.com/js-temporal/temporal-polyfill/issues/44
+        if (difference > -1) {
+            value = _month.add({months: difference}).toString();
+            return;
+        }
+
+        // UNKNOWN: Will this have issues with year 0 being subtracted into year -1 (1 BC)?
+        // Probably, but who is crazy enough to try subtracting all that way via the UI :)?
+        value = Temporal.PlainYearMonth.from({
+            calendar: _month.calendar,
+            year: _month.month > 1 ? _month.year : _month.year - 1,
+            month: _month.month > 1 ? _month.month - 1 : _month.monthsInYear,
+        }).toString({calendarName: "always"});
     }
 
-    $: _year = Temporal.PlainYearMonth.from(value);
+    $: _month = Temporal.PlainYearMonth.from(value);
 </script>
 
-<PickerContainer {...$$props} bind:element class="year-selector {_class}">
+<PickerContainer {...$$props} bind:element class="month-stepper {_class}">
     <PickerSection variation="flex">
         <PickerHeader>
-            {_year.toLocaleString(locale, {year: "numeric"})}
+            {_month.toLocaleString(locale, {month: "long", year: "numeric"})}
         </PickerHeader>
 
         <Spacer variation="inline" />
 
         <PickerButton
-            disabled={!is_year_in_range(_year, undefined, min)}
+            disabled={!is_month_in_range(_month, undefined, min)}
             {palette}
-            on:click={on_year_select.bind(null, -1)}
+            on:click={on_month_select.bind(null, -1)}
         >
             <slot name="previous">&lt;</slot>
         </PickerButton>
 
         <PickerButton
-            disabled={!is_year_in_range(_year, max)}
+            disabled={!is_month_in_range(_month, max)}
             {palette}
-            on:click={on_year_select.bind(null, 1)}
+            on:click={on_month_select.bind(null, 1)}
         >
             <slot name="next">&gt;</slot>
         </PickerButton>
