+++ src/lib/components/pickers/monthstepper/index.ts
@@ -1,0 +1,1 @@
+export {default as MonthStepper} from "./MonthStepper.svelte";
