--- src/lib/components/pickers/yearselector/index.ts
@@ -1,1 +1,0 @@
-export {default as YearSelector} from "./YearSelector.svelte";
