+++ src/lib/components/pickers/yearstepper/index.ts
@@ -1,0 +1,1 @@
+export {default as YearStepper} from "./YearStepper.svelte";
